+++ src/aioway/attrs/devices.py
