+++ src/aioway/attrs/shapes.py
