+++ src/aioway/codecs/images.py
@@ -11,7 +11,7 @@ from torchvision import io as vio
 from torchvision.transforms import v2 as tt
 
 from aioway._utils import current_fake_mode, torch_set_fake_mode_func
-from aioway.spaces import Attr, AttrLike
+from aioway.attrs import Attr, AttrLike
 
 from ._bases import TorchCompatible
