+++ src/aioway/codecs/videos.py
@@ -9,7 +9,7 @@ import torch
 from torchcodec import decoders as dec
 
 from aioway._utils import current_fake_mode, num_threads, torch_set_fake_mode_func
-from aioway.spaces import Attr
+from aioway.attrs import Attr
 
 from ._av import VideoStream
 from ._bases import TorchCompatible
