+++ src/aioway/spaces/__init__.py
@@ -1,6 +1,5 @@
 # Copyright (c) AIoWay Authors - All Rights Reserved
 
-from .attrs import *
 from .dims import *
 from .exact import *
 from .gym import *
