+++ tests/spaces/attrs/test_devices.py
@@ -3,7 +3,7 @@
 import pytest
 import torch
 
-from aioway.spaces import Device
+from aioway.attrs import Device
 
 
 def _cpus():
