+++ tests/spaces/attrs/test_dtypes.py
@@ -7,7 +7,7 @@ import numpy as np
 import pytest
 import torch
 
-from aioway.spaces import DType, DTypeFamily
+from aioway.attrs import DType, DTypeFamily
 
 
 @dcls.dataclass(frozen=True)
