+++ tests/test_emitters.py
@@ -7,10 +7,11 @@ from torch import optim
 
 from aioway._iters import TensorIter
 from aioway._ufuncs import BuiltUFunc, UFunc, UFuncThunk
+from aioway.attrs import Attr, Shape
 from aioway.emits import MlpEmitter, emit, linear_shape
 from aioway.io import TensorListIter, TensorStream
 from aioway.relalg import LoaderOpt
-from aioway.spaces import Attr, AttrSpace, Shape, ShapeSpace
+from aioway.spaces import AttrSpace, ShapeSpace
 from aioway.torch.nn import Linear, MSELoss, NnUFunc
 from aioway.torch.optim import OptimizerUFunc
