+++ tests/test_indices.py
@@ -4,9 +4,9 @@ import pytest
 import torch
 
 from aioway._iters import TensorIter
+from aioway.attrs import DType
 from aioway.indices import AnnIndexIter, AnnIndexTrainerIter, FaissIndex
 from aioway.io import TensorListIter
-from aioway.spaces import DType
 
 
 @pytest.fixture
